fix: use larger datatype to store key representation

+++ src/model/crdt/element.h
@@ -4,18 +4,18 @@
 #include <stdbool.h>
 
 typedef struct {
-  unsigned int key;      // bit representation of the key.
-  unsigned int user_ids; // user ids associated with each key token.
-  unsigned int depth;    // length of the key.
-  unsigned int version;  // the version of the element.
-  void* value;           // value of the element.
+  unsigned long key;      // bit representation of the key, up to 6 tokens.
+  unsigned long user_ids; // user ids associated with each key token.
+  int depth;              // length of the key.
+  unsigned int version;   // the version of the element.
+  void* value;            // value of the element.
 } element;
 
-unsigned int key_from_tokens(unsigned int depth, ...);
-unsigned int user_ids_from_ids(unsigned int depth, ...);
+unsigned long key_from_tokens(int depth, ...);
+unsigned long user_ids_from_ids(int depth, ...);
 int key_compare(element* l, element* r);
 bool key_equal(element* l, element* r);
-void set_key_tokens(element* e, unsigned int depth, ...);
+void set_key_tokens(element* e, int depth, ...);
 
 #endif
